fix: flavorNote로 변경 및 likeCount Integer로 변경

+++ backend/coffeeing/src/main/java/com/ssafy/coffeeing/modules/feed/dto/FeedDetailResponse.java
@@ -10,7 +10,7 @@ public record FeedDetailResponse(
         String content,
         Tag tag,
         Long registerId,
-        Long likeCount,
+        Integer likeCount,
         String registerName,
         String registerProfileImg,
         Boolean isLike,
